Web-based Spoon ver. 0.1
This is an attempt to run Spoon as a web app.
Spoon relies on SWT for UI widgets, which is great for being OS agnostic, but it only runs as a desktop app.
RWT provides web UIs with SWT API, so replacing SWT with RWT allows Spoon to run as a web app with a little code change.
Having said that, some APIs are not implemented; hence, a little more code change is required than it sounds.

These are the major changes so far:
 - Add org.pentaho.di.ui.spoon.BasicApplication, which configures web app
 - Modify ui/ivy.xml in order to add RWT-related dependencies and remove SWT
 - Many comment-outs (not complete yet!) to avoid a compile error due to RWT/SWT difference